+++ CHANGELOG.md
@@ -566,3 +566,4 @@
 * Remove `Pointable` typeclass
 * Rename `point_` to `point`
 * Define `extract_` method
+* Remove `Extractable` typeclass